pyproject.toml
[project]
name = "python-playground"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "google-genai",
    "ragas",
    "fastmcp",
    "httpx",
    "uvicorn",
    "fastapi",
]
